Removed scenario attack folders from c/examples to remove code duplication. Moved the attack testing code into scenario_example.
Updated scenario_example/client.cpp to add the attack testing functionality using the csv file.
