--- examples/scenario_attack1_seqnum/client.cpp
@@ -1,108 +1,0 @@
-extern "C" {
-#include <c/c_api.h>
-#include <c/ipfs.h>
-}
-
-#include <unistd.h>
-
-#include <fstream>
-#include <iostream>
-#include <thread>
-
-enum class AttackType {
-    NONE,
-    REPLAY
-};
-
-static AttackType parseAttackType(const std::string& s) {
-    if (s == "REPLAY") return AttackType::REPLAY;
-    return AttackType::NONE;
-}
-
-int main(int argc, char *argv[]) {
-    if (argc != 3) {
-        std::cerr << "Usage: " << argv[0] << " <config_path> <csv_file_path>"
-                  << std::endl;
-        exit(1);
-    }
-
-    // Standard SST initialization
-    char *config_path = argv[1];
-    SST_ctx_t *ctx = init_SST(config_path);
-
-    session_key_list_t *s_key_list = get_session_key(ctx, NULL);
-    if (s_key_list == NULL) {
-        std::cerr << "Failed to get session key. Returning NULL.\n"
-                  << ::std::endl;
-        exit(1);
-    }
-
-    // SST Connect to server
-    SST_session_ctx_t *session_ctx =
-        secure_connect_to_server(&s_key_list->s_key[0], ctx);
-
-    // Read CSV file
-    std::string my_file_path = argv[2];
-    std::ifstream file(my_file_path);
-    std::string line;
-
-    // Because client sends messages, changing the received sequence number has no effect
-    // session_ctx->received_seq_num = 10; // Attack the received sequence number
-    session_ctx->sent_seq_num = -7; // Attack the sent sequence number
-
-    while (std::getline(file, line)) {
-        //  split the line into 4 tokens
-        size_t comma1 = line.find(',');
-        size_t comma2 = line.find(',', comma1 + 1);
-        size_t comma3 = line.find(',', comma2 + 1);
-
-        // Sleep for the specified time
-        std::string sleep_time_str = line.substr(0, comma1);
-        int sleep_time = std::stoi(sleep_time_str);
-        std::this_thread::sleep_for(std::chrono::milliseconds(sleep_time));
-
-        // Send message to server
-        std::string message = line.substr(comma1 + 1, comma2 - comma1 - 1);
-        send_secure_message(const_cast<char *>(message.c_str()),
-                            message.length(), session_ctx);
-
-        // Parse the attack type
-        std::string attack_type_str = (comma2 != std::string::npos)
-        ? line.substr(comma2 + 1, (comma3 == std::string::npos ? std::string::npos : comma3 - comma2 - 1)) // if there is a 3rd column, grab it
-        : ""; // else, use the empty string
-
-        AttackType atype = parseAttackType(attack_type_str);
-
-        // Optional: parameter for the attack type if applicable
-        std::string attack_param = (comma3 != std::string::npos)
-        ? line.substr(comma3 + 1)
-        : "";
-
-        switch (atype) {
-            case AttackType::REPLAY:
-                if      (attack_param == "seq--") {
-                    session_ctx->sent_seq_num--;
-                }
-                else if (attack_param == "seq++") {
-                    session_ctx->sent_seq_num++;
-                }
-                else if (attack_param.rfind("seq=", 0) == 0) {
-                    // parse “seq=#”
-                    int v = std::stoi(attack_param.substr(4));
-                    session_ctx->sent_seq_num = v;
-                }
-                break;
-
-            case AttackType::NONE:
-            default:
-                break;
-            }
-
-    }
-
-    free(session_ctx);
-    free_session_key_list_t(s_key_list);
-    free_SST_ctx_t(ctx);
-
-    return EXIT_SUCCESS;
-}
