--- examples/scenario_attack1_seqnum/example_messages.csv
@@ -1,3 +1,0 @@
-1000,Hello
-500,World
-1500,!
