--- examples/scenario_attacks/example_messages.csv
@@ -1,3 +1,0 @@
-1000,Hello,dos,100000
-500,World
-1500,!
